multiNotification/requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.6
discord-webhook==0.14.0
idna==3.2
pydantic==1.8.2
python-telegram-bot==13.7
pytz==2021.1
requests==2.26.0
six==1.16.0
tornado==6.1
typing-extensions==3.10.0.2
tzlocal==3.0
urllib3==1.26.7
